Messaging and Threads (#6)
Added a data model for Thread which owns a thread of messages for a group of users. Also, changed the Message data model to belong to a thread instead of a to user.

I implemented the following endpoints:

POST /messages/send: takes the message data and saves it to storage. Call with something like:
{
	"sender": "24d6d9cf-8435-41c0-9417-5eeccdd8fb35",
	"thread": "e3d4ca7f-4175-4bcf-818d-43e7cd41d4c8",
	"payload": "yo"
}
GET /messages/get/<threadId> (probably should change this to just /messages/<threadId>): returns all of the messages for the thread.
GET /threads/<userId>: returns all of the threads that a user is in.
All of these endpoints require that you have a header called Cookie with a value like sid=e3d4ca7f-4175-4bcf-818d-43e7cd41d4c8. This will be the auth token that our session management system will generate in the future. For the time being, it doesn't matter what this is,
as long as it's a valid UUID.