Basic authentication flow (#13)
Add the ability to perform basic authentication using facebook.

Go to localhost:8080/login and you will be redirected to fb. From here we get an authorization code that can be used to get an access token later on to deal with user data.

Currently we do not get this access token so thats a big TODO.

Also contains a basic secrets manager which allows you to load secrets from a json file in the secrets directory. Not sure what the best way to distribute this file is since it should remain secure, for now its just local to my computer. but we can maybe store this secret data in a db in the future.